Refactor project structure and styles: remove unused components, update imports, and enhance styling for better organization and maintainability; add prop-types dependency for type checking.
