+++ package.json
@@ -18,6 +18,7 @@
     "@fortawesome/react-fontawesome": "^0.2.2",
     "font-awesome": "^4.7.0",
     "i": "^0.3.7",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "styled-components": "^6.1.17"
