+++ src/App.jsx
@@ -1,11 +1,9 @@
 import "./index.css";
-
 import { AboutMe } from "./Sections/MainBody/About/AboutMe.jsx";
 import { Header } from "./Sections/Header/Head.jsx";
 import { Projects } from "./Sections/MainBody/Projects/Projects.jsx";
-import { ToolkitSection } from "./Sections/Toolkit.jsx";
-
-import { Skills } from "./Sections/Skills.jsx";
+import { ToolkitSection } from "./Sections/MainBody/Tools/Toolkit.jsx";
+import { Skills } from "./Sections/MainBody/Skills/Skills.jsx";
 import { Footer } from "./Sections/Footer/FooterContent.jsx";
 import { Updates } from "./Sections/MainBody/Updates/Updates.jsx";
