+++ src/index.css
@@ -117,11 +117,6 @@ main {
   width: 100dvw;
 }
 
-.project-icons > path,
-.footer-contacts > ul > li svg path {
-  fill: var(--color-accent);
-}
-
 @media (min-width: 768px) {
   body {
     font-size: 1.2rem;
@@ -132,7 +127,7 @@ main {
   }
 
   h3 {
-    font-size: 3rem;
+    font-size: 2rem;
   }
   H4 {
     font-size: 1.5rem;
@@ -150,11 +145,11 @@ main {
   }
 
   h2 {
-    font-size: 4rem;
+    font-size: 3rem;
   }
 
   h3 {
-    font-size: 2.5rem;
+    font-size: 2rem;
   }
   H4 {
     font-size: 1.5rem;
